+++ tsMuxer/simplePacketizerReader.cpp
@@ -31,16 +31,15 @@ void SimplePacketizerReader::doMplsCorrection()
 {
     if (m_curMplsIndex == -1)
         return;
-    if (m_curPts >= (m_lastMplsTime - mplsEps) && m_curMplsIndex < m_mplsInfo.size() - 1)
+    if (m_curPts >= (m_lastMplsTime - mplsEps) && m_curMplsIndex < static_cast<int>(m_mplsInfo.size() - 1))
     {
         m_curMplsIndex++;
         if (m_mplsInfo[m_curMplsIndex].connection_condition == 5)
         {
             m_mplsOffset += m_curPts - m_lastMplsTime;
-            // m_curPts = m_lastMplsTime; // fix PTS up
         }
-        m_lastMplsTime += static_cast<int64_t>(
-            (m_mplsInfo[m_curMplsIndex].OUT_time - m_mplsInfo[m_curMplsIndex].IN_time) * (INTERNAL_PTS_FREQ / 45000.0));
+        m_lastMplsTime +=
+            (m_mplsInfo[m_curMplsIndex].OUT_time - m_mplsInfo[m_curMplsIndex].IN_time) * (INTERNAL_PTS_FREQ / 45000.0);
     }
 }
 
@@ -172,7 +171,7 @@ int SimplePacketizerReader::readPacket(AVPacket& avPacket)
             LTRACE(LT_INFO, 2,
                    getCodecInfo().displayName
                        << " stream (track " << m_streamIndex << "): bad frame detected at position"
-                       << floatToTime((avPacket.pts - PTS_CONST_OFFSET) / (double)INTERNAL_PTS_FREQ, ',')
+                       << floatToTime((double)(avPacket.pts - PTS_CONST_OFFSET) / INTERNAL_PTS_FREQ, ',')
                        << ". Resync stream.");
             m_needSync = true;
             return 0;
@@ -214,7 +213,7 @@ int SimplePacketizerReader::readPacket(AVPacket& avPacket)
                     LTRACE(LT_INFO, 2,
                            getCodecInfo().displayName
                                << " stream (track " << m_streamIndex << "): overlapped frame detected at position "
-                               << floatToTime((avPacket.pts - PTS_CONST_OFFSET) / (double)INTERNAL_PTS_FREQ, ',')
+                            << floatToTime((double)(avPacket.pts - PTS_CONST_OFFSET) / INTERNAL_PTS_FREQ, ',')
                                << ". Remove frame.");
                 m_mplsOffset -= getFrameDuration();
                 m_curPts -= getFrameDuration();
